some more updates on the interface implementation of the Data class
git-svn-id: https://tulip-plugins.googlecode.com/svn/trunk@43 4ddaf586-9532-11dd-8e9d-3bce277bc375

+++ miniprojet/IData.h
@@ -11,12 +11,12 @@
 template<typename T> class IData
 {
 	public:
-		virtual int size() const = 0;
+		virtual size_t size() const = 0;
 		virtual void add(T, T) = 0;
 		virtual T getX(int) const = 0;
 		virtual T getY(int) const = 0;
-		virtual void setX(int, T) = 0;
-		virtual void setY(int, T) = 0;
+		virtual void setX(size_t, T) = 0;
+		virtual void setY(size_t, T) = 0;
 };
 
 #endif /* IDATA_H_ */
